requirements.txt
apache-airflow-providers-google
pyarrow
joblib
sklearn
